Move not API classes to an "internal" package
Isolating internal classes into dedicated packages makes easier to create consistent APIs through Java 9's modules.
